Get Inngest working with API request
